agents/prefetch/requirements.txt
# Dependencies for pre-fetch scripts
# All are already available in the project environment or are stdlib.
requests>=2.28.0
python-dotenv>=1.0.0
